hostapp-update-hooks: remove 99-resin-uboot hook
This was added in the hooks because the jetsons use a different hook
as their partition numbers seem to be in hex.

https://github.com/balena-os/meta-balena/pull/1349 adds this in
meta-balena. Remove the custom hook here as there are updates to this
hook related to rollbacks in meta-balena that have not reached the
jetson family.

Change-type: patch
Changelog-entry: Remove custom device u-boot hook to let the meta-balena one work
Signed-off-by: Zubair Lutfullah Kakakhel <zubair@balena.io>